refactor to buffer
